WIP: reduced width of app win wide window
created a size break for small vs large width windows
