+++ web/src/components/codecarto/codecarto.css
@@ -11,31 +11,39 @@
   --c-file: var(--c-secondary);
   --c-font: #e1c48f;
   --c-graph: #705d50;
-  --cb-directory: var(--c-primary);
+  --cb-directory: #3a2d23;
   --cb-folder-active: #433920;
   --cb-folder-hover: var(--c-secondary-light);
   --cb-folder: var(--c-secondary);
-  --cb-file-disabled: #8d8374;
+  --cb-file-disabled: #9d9282;
   --cb-file-active: var(--c-secondary-light);
   --cb-file-hover: var(--c-secondary-light);
-  --cb-file: var(--c-primary);
+  --cb-file: var(--cb-directory);
   --w-nav-left: 5rem;
-  --z-nav: 1;
-  --z-plot: 2;
+  --z-plot: 1;
+  --z-nav: 2;
 }
 
 .app {
+  height: calc(100vh - 20px) !important;
+}
+
+.codecarto {
   display: grid;
   grid-template-rows: auto auto 1fr;
   flex-direction: column;
-  height: calc(100vh - 20px) !important;
+  height: 100%;
+  padding-inline: 10%;
   font-size: 20px;
 }
 
 .header {
-  font-size: 30px;
   font-weight: bold;
   margin-bottom: 5px;
+}
+
+.app_header {
+  font-size: 30px;
 }
 
 .parser_graph_desc {
@@ -44,25 +52,37 @@
 }
 
 /* ###########  DIRECTORY NAV TOGGLE  ######### */
-/* .nav {
-  position: fixed;
-  margin-top: 0;
-  top: 0;
-  left: 0;
+.nav {
+  position: absolute;
+  height: calc(100vh - 10.1em) !important;
+  top: 9.3em;
+  border-radius: 5px;
   z-index: var(--z-nav);
 }
 .nav__toggle_btn {
+  position: relative;
   width: 1em;
+  margin-left: 0;
   border-top-left-radius: 0;
   border-bottom-left-radius: 0;
 }
 .nav__toggle_btn--open {
   width: 2em;
-  margin-left: calc(18.8em + 0.3em);
-  border-top-left-radius: var(--border-radius);
-  border-bottom-left-radius: var(--border-radius);
+  margin-left: calc(14.8em);
 }
 .nav__content {
+  position: absolute;
+  height: 100%;
   width: 15em;
-  margin-left: calc(-15em - 2.1em);
-} */
+  margin-left: -15em;
+}
+.nav__toggle_btn--open + .nav__content {
+  margin-left: 0;
+}
+
+/* check if vw is of a certain width */
+@media (max-width: 700px) {
+  .codecarto {
+    padding-inline: 0;
+  }
+}
