+++ web/src/components/codecarto/codecarto.ts
@@ -1,15 +1,15 @@
 import m from 'mithril';
 
 import { ICell } from '../../state';
-import { Plot } from '../plot/plot';
-import { UrlInput } from '../url_input/url_input';
-import { Directory } from '../directory/directory';
 import { Nav } from '../navigation/nav';
+import { UrlInput } from '../url_input/url_input';
+import { Plot } from '../plot/plot';
+import { Directory } from '../directory/directory';
 import { handleGithubURL, plotGithubUrl } from '../../services/repo_service';
 import './codecarto.css';
 
 export const CodeCarto = (cell: ICell) => {
-  const title = m('div.header', 'Code Cartographer');
+  const title = m('div.header.app_header', 'Code Cartographer');
 
   const handleUrlInput = async () => {
     handleGithubURL(cell, updateGithubData);
@@ -30,40 +30,38 @@ export const CodeCarto = (cell: ICell) => {
     m.redraw();
   };
 
-  // const setSelectedUrl = (url: string) => {
-  //   cell.state.selected_file_url = url;
-  //   plotGithubUrl(cell, handlePlotData);
-  // };
+  const setSelectedFile = (url: string) => {
+    cell.state.selected_file_url = url;
+    plotGithubUrl(cell, handlePlotData);
+  };
 
-  // const handlePlotData = (data: Array<object>) => {
-  //   let nbFrame: m.Vnode[] = [];
-  //   if (data && data.length > 0) {
-  //     // Create an iframe for each output
-  //     data.forEach((output) => {
-  //       if (output['text/html']) {
-  //         nbFrame.push(
-  //           m('iframe.graph_content.nbFrame', {
-  //             srcdoc: output['text/html'],
-  //           })
-  //         );
-  //       }
-  //     });
-  //   }
+  const handlePlotData = (data: Array<object>) => {
+    let nbFrame: m.Vnode[] = [];
+    if (data && data.length > 0) {
+      // Create an iframe for each output
+      data.forEach((output) => {
+        if (output['text/html']) {
+          nbFrame.push(
+            m('iframe.graph_content.nbFrame', {
+              srcdoc: output['text/html'],
+            })
+          );
+        }
+      });
+    }
 
-  //   // Update the cell with the new content
-  //   cell.update({
-  //     graph_content: nbFrame,
-  //     showContentNav: false,
-  //   });
+    // Update the cell with the new content
+    cell.update({
+      graph_content: nbFrame,
+      showContentNav: false,
+    });
 
-  //   // Trigger a redraw to update the view
-  //   m.redraw();
-  // };
+    // Trigger a redraw to update the view
+    m.redraw();
+  };
 
   return [
-    title,
-    UrlInput(cell, handleUrlInput),
-    //Nav(cell, 'showContentNav', Directory(cell, setSelectedUrl)),
-    Plot(cell),
+    Nav(cell, 'showContentNav', Directory(cell, setSelectedFile)),
+    m('div.codecarto', [title, UrlInput(cell, handleUrlInput), Plot(cell)]),
   ];
 };
