+++ web/src/components/plot/plot.css
@@ -1,6 +1,6 @@
 .plot {
-  padding: 5px;
   margin-top: 8px;
+  width: 100%;
   background-color: var(--c-primary-light);
   justify-content: center;
   align-items: center;
@@ -9,15 +9,14 @@
 }
 
 .graph {
-  position: absolute;
-  height: calc(100vh - 8.8em) !important;
-  width: calc(100vw - 1.35em) !important;
+  height: 100%;
+  width: 100%;
   border-radius: 5px;
 }
 
 .graph_content {
-  height: 100% !important;
-  width: 100% !important;
+  height: 100%;
+  width: 100%;
   border: none;
   overflow: auto;
 }
